+++ src/Http/Middleware/ContentLanguageMiddleware.php
@@ -5,6 +5,7 @@ namespace Lambdadigamma\LaravelApiLanguage\Http\Middleware;
 use Closure;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\App;
+use Symfony\Component\HttpFoundation\Response;
 
 class ContentLanguageMiddleware
 {
@@ -13,8 +14,8 @@ class ContentLanguageMiddleware
         $response = $next($request);
         $contentLanguage = App::getLocale() ?? config('app.fallback_locale') ?? 'en';
 
-        if (method_exists($response, 'header')) {
-            $response->header('Content-Language', $contentLanguage);
+        if ($response instanceof Response && ! $response->headers->has('Content-Language')) {
+            $response->headers->set('Content-Language', $contentLanguage);
         }
 
         return $response;